notewise-backend-v2/requirements.txt
fastapi=0.129.2
pydantic=2.12.5
uvicorn=0.41.0
sqlalchemy=2.0.46
python-dotenv=1.2.1
alembic=1.18.4 
pydantic_settings=2.13.1
psycopg2-binary=2.9.11
python-multipart=0.0.22
azure-storage-blob
python-jose[cryptography]>=3.3.0 
httpx>=0.27.0 